fix rich progress metric render on epoch end
