+++ tests/callbacks/test_tqdm_progress_bar.py
@@ -14,7 +14,6 @@
 import os
 import pickle
 import sys
-from collections import defaultdict
 from typing import Union
 from unittest import mock
 from unittest.mock import ANY, call
@@ -556,65 +555,3 @@ def test_get_progress_bar_metrics(tmpdir: str):
     assert "loss" in standard_metrics.keys()
     assert "split_idx" in standard_metrics.keys()
     assert "v_num" not in standard_metrics.keys()
-
-
-def test_tqdm_progress_bar_correct_value_epoch_end(tmpdir):
-    class MockedProgressBar(TQDMProgressBar):
-        calls = defaultdict(list)
-
-        def get_metrics(self, trainer, pl_module):
-            items = super().get_metrics(trainer, model)
-            del items["v_num"]
-            del items["loss"]
-            # this is equivalent to mocking `set_postfix` as this method gets called every time
-            self.calls[trainer.state.fn].append(
-                (trainer.state.stage, trainer.current_epoch, trainer.global_step, items)
-            )
-            return items
-
-    class MyModel(BoringModel):
-        def training_step(self, batch, batch_idx):
-            self.log("a", self.global_step, prog_bar=True, on_step=False, on_epoch=True, reduce_fx=max)
-            return super().training_step(batch, batch_idx)
-
-        def validation_step(self, batch, batch_idx):
-            self.log("b", self.global_step, prog_bar=True, on_step=False, on_epoch=True, reduce_fx=max)
-            return super().validation_step(batch, batch_idx)
-
-        def test_step(self, batch, batch_idx):
-            self.log("c", self.global_step, prog_bar=True, on_step=False, on_epoch=True, reduce_fx=max)
-            return super().test_step(batch, batch_idx)
-
-    model = MyModel()
-    pbar = MockedProgressBar()
-    trainer = Trainer(
-        default_root_dir=tmpdir,
-        limit_train_batches=2,
-        limit_val_batches=2,
-        limit_test_batches=2,
-        max_epochs=2,
-        enable_model_summary=False,
-        enable_checkpointing=False,
-        log_every_n_steps=1,
-        callbacks=pbar,
-    )
-
-    trainer.fit(model)
-    assert pbar.calls["fit"] == [
-        ("sanity_check", 0, 0, {"b": 0}),
-        ("train", 0, 0, {}),
-        ("train", 0, 1, {}),
-        ("validate", 0, 1, {"b": 1}),  # validation end
-        # epoch end over, `on_epoch=True` metrics are computed
-        ("train", 0, 2, {"a": 1, "b": 1}),  # training epoch end
-        ("train", 1, 2, {"a": 1, "b": 1}),
-        ("train", 1, 3, {"a": 1, "b": 1}),
-        ("validate", 1, 3, {"a": 1, "b": 3}),  # validation end
-        ("train", 1, 4, {"a": 3, "b": 3}),  # training epoch end
-    ]
-
-    trainer.validate(model, verbose=False)
-    assert pbar.calls["validate"] == []
-
-    trainer.test(model, verbose=False)
-    assert pbar.calls["test"] == []
